:recycle:refactor: 패키지 구조 변경 및 도메인 변경
