--- src/main/java/com/club_board/club_board_server/domain/Club.java
@@ -1,14 +1,0 @@
-package com.club_board.club_board_server.domain;
-import jakarta.persistence.*;
-
-@Entity
-
-public class Club {
-
-    @Id @GeneratedValue
-    @Column(name="club_id")
-    private Long id;
-
-    @Column(name="club_name")
-    private String name;
-}
