--- src/main/java/com/club_board/club_board_server/domain/Role.java
@@ -1,7 +1,0 @@
-package com.club_board.club_board_server.domain;
-
-public enum Role {
-    OWNER,
-    ADMIN,
-    USER
-}
