+++ src/main/java/com/club_board/club_board_server/domain/user/Role.java
@@ -1,0 +1,7 @@
+package com.club_board.club_board_server.domain.user;
+
+public enum Role {
+    OWNER,
+    ADMIN,
+    USER
+}
