+++ src/main/java/com/club_board/club_board_server/domain/user/User.java
@@ -1,11 +1,10 @@
-package com.club_board.club_board_server.domain;
+package com.club_board.club_board_server.domain.user;
 import jakarta.persistence.*;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import java.time.LocalDate;
-import java.util.HashSet;
-import java.util.Set;
+
 @NoArgsConstructor
 @Entity
 @Getter
