+++ src/main/java/com/club_board/club_board_server/repository/user/UserRepository.java
@@ -1,5 +1,5 @@
-package com.club_board.club_board_server.repository;
-import com.club_board.club_board_server.domain.User;
+package com.club_board.club_board_server.repository.user;
+import com.club_board.club_board_server.domain.user.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import java.util.Optional;
