+++ src/main/java/com/club_board/club_board_server/service/auth/CustomUserDetailsService.java
@@ -1,7 +1,7 @@
 package com.club_board.club_board_server.service.auth;
-import com.club_board.club_board_server.domain.CustomUserDetails;
-import com.club_board.club_board_server.domain.User;
-import com.club_board.club_board_server.repository.UserRepository;
+import com.club_board.club_board_server.domain.user.CustomUserDetails;
+import com.club_board.club_board_server.domain.user.User;
+import com.club_board.club_board_server.repository.user.UserRepository;
 import com.club_board.club_board_server.response.exception.BusinessException;
 import com.club_board.club_board_server.response.exception.ExceptionType;
 import lombok.RequiredArgsConstructor;
